feat: allow importing Mongoose without importing mongodb driver
Re: #4292
Re: #3774
Re: #12052
Re: #10970
