+++ lib/query.js
@@ -2260,7 +2260,7 @@ Query.prototype._find = async function _find() {
   const filter = this._conditions;
   const fields = options.projection;
 
-  const cursor = await this._collection.collection.find(filter, options);
+  const cursor = await this.mongooseCollection.find(filter, options);
   if (options.explain) {
     return cursor.explain();
   }
@@ -2509,7 +2509,7 @@ Query.prototype._findOne = async function _findOne() {
   this._applyTranslateAliases(options);
 
   // don't pass in the conditions because we already merged them in
-  const doc = await this._collection.collection.findOne(this._conditions, options);
+  const doc = await this.mongooseCollection.findOne(this._conditions, options);
   return new Promise((resolve, reject) => {
     this._completeOne(doc, null, _wrapThunkCallback(this, (err, res) => {
       if (err) {
@@ -2611,7 +2611,7 @@ Query.prototype._count = async function _count() {
 
   const conds = this._conditions;
 
-  return this._collection.collection.count(conds, options);
+  return this.mongooseCollection.count(conds, options);
 };
 
 /**
@@ -2641,7 +2641,7 @@ Query.prototype._countDocuments = async function _countDocuments() {
 
   const conds = this._conditions;
 
-  return this._collection.collection.countDocuments(conds, options);
+  return this.mongooseCollection.countDocuments(conds, options);
 };
 
 /*!
@@ -2686,7 +2686,7 @@ Query.prototype._estimatedDocumentCount = async function _estimatedDocumentCount
 
   const options = this._optionsForExec();
 
-  return this._collection.collection.estimatedDocumentCount(options);
+  return this.mongooseCollection.estimatedDocumentCount(options);
 };
 
 /**
@@ -2855,7 +2855,7 @@ Query.prototype.__distinct = async function __distinct() {
   const options = this._optionsForExec();
   this._applyTranslateAliases(options);
 
-  return this._collection.collection.
+  return this.mongooseCollection.
     distinct(this._distinct, this._conditions, options);
 };
 
@@ -3015,7 +3015,7 @@ Query.prototype._deleteOne = async function _deleteOne() {
   const options = this._optionsForExec();
   this._applyTranslateAliases(options);
 
-  return this._collection.collection.deleteOne(this._conditions, options);
+  return this.mongooseCollection.deleteOne(this._conditions, options);
 };
 
 /**
@@ -3093,7 +3093,7 @@ Query.prototype._deleteMany = async function _deleteMany() {
   const options = this._optionsForExec();
   this._applyTranslateAliases(options);
 
-  return this._collection.collection.deleteMany(this._conditions, options);
+  return this.mongooseCollection.deleteMany(this._conditions, options);
 };
 
 /**
@@ -3338,7 +3338,7 @@ Query.prototype._findOneAndUpdate = async function _findOneAndUpdate() {
     this._update = this._update.toBSON();
   }
 
-  let res = await this._collection.collection.findOneAndUpdate(this._conditions, this._update, options);
+  let res = await this.mongooseCollection.findOneAndUpdate(this._conditions, this._update, options);
   for (const fn of this._transforms) {
     res = fn(res);
   }
@@ -3487,7 +3487,7 @@ Query.prototype._findOneAndDelete = async function _findOneAndDelete() {
   const options = this._optionsForExec(this.model);
   this._applyTranslateAliases(options);
 
-  let res = await this._collection.collection.findOneAndDelete(filter, options);
+  let res = await this.mongooseCollection.findOneAndDelete(filter, options);
   for (const fn of this._transforms) {
     res = fn(res);
   }
@@ -3623,7 +3623,7 @@ Query.prototype._findOneAndReplace = async function _findOneAndReplace() {
       throw validationError;
     }
 
-    let res = await this._collection.collection.findOneAndReplace(filter, this._update || {}, options);
+    let res = await this.mongooseCollection.findOneAndReplace(filter, this._update || {}, options);
 
     for (const fn of this._transforms) {
       res = fn(res);
@@ -3649,7 +3649,7 @@ Query.prototype._findOneAndReplace = async function _findOneAndReplace() {
     castedDoc = castedDoc.toBSON();
   }
 
-  let res = await this._collection.collection.findOneAndReplace(filter, castedDoc, options);
+  let res = await this.mongooseCollection.findOneAndReplace(filter, castedDoc, options);
 
   for (const fn of this._transforms) {
     res = fn(res);
@@ -3876,7 +3876,7 @@ async function _updateThunk(op) {
     this._update = this._update.toBSON();
   }
 
-  return this._collection.collection[op](castedQuery, this._update, options);
+  return this.mongooseCollection[op](castedQuery, this._update, options);
 }
 
 /**
